backend/requirements.txt
Flask
tensorflow-cpu
numpy
Pillow
firebase-admin
requests
gunicorn